+++ src/common/outbound.h
@@ -25,7 +25,7 @@
 extern const struct commands xc_cmds[];
 extern GSList *menu_list;
 
-int auto_insert (char *dest, int destlen, unsigned char *src, char *word[], char *word_eol[],
+int auto_insert (char *dest, gsize destlen, unsigned char *src, char *word[], char *word_eol[],
 				 char *a, char *c, char *d, char *e, char *h, char *n, char *s, char *u);
 char *command_insert_vars (session *sess, char *cmd);
 int handle_command (session *sess, char *cmd, int check_spch);
